+++ lib/new_relic/agent/instrumentation/sinatra.rb
@@ -45,6 +45,11 @@ DependencyDetection.defer do
 
       ::Sinatra::Base.class_eval do
         class << self
+          def build_with_newrelic(*args, &block)
+            build_with_tracing(*args) do 
+              build_without_newrelic(*args, &block)
+            end
+          end
           alias build_without_newrelic build
           alias build build_with_newrelic
         end
